requirements.txt
selenium
allure-behave
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pyTelegramBotAPI
telethon
gspread
antigate
pytz
2captcha-python
bs4
TwoCaptcha
configparser
requests
beautifulsoup4
behave
lxml
undetected-chromedriver
selenium-wire
Pillow
httpx[http2]
oauth2client
nord
clipboard
pydrive
python-dateutil
webdriver-manager